reset warningCounter when restarting dxgrabber
when switching to / from exclusive fullscreen mode, it will fail to capture a couple of frames, and usually the dxgrabber will restart because of DXGI_ERROR_ACCESS_LOST (this is normal when going to exclusive fullscreen). However, if you do this multiple times eventually the warningCounter is reduced to 0 and the grabber will be disabled. I think it's better to reset the count so it keeps working even though it has been restarted multiple times. I've also increased the warning counter to 15 so it's less likely to disable the grabber.